+++ app/src/main/java/com/example/nb2998/taskwendor/Models/SingleItem.java
@@ -3,9 +3,9 @@ package com.example.nb2998.taskwendor.Models;
 public class SingleItem {
     private String name, image_url;
     private int tot_units, left_units;
-    private float price;
+    private int price;
 
-    public SingleItem(String name, String image_url, int tot_units, int left_units, float price) {
+    public SingleItem(String name, String image_url, int tot_units, int left_units, int price) {
         this.name = name;
         this.image_url = image_url;
         this.tot_units = tot_units;
